requirements.txt
blinker
dnspython
email-validator
flask
flask_login
flask_sqlalchemy
flask_wtf
wtforms
pydantic
flask_pydantic
email-validator
